refactor: put components in folder
