+++ react-start/src/pages/about-us/about-us.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import About from '../components/about';
+import About from '../../components/about/about';
 
 class AboutUs extends React.Component {
   render(): React.ReactNode {
