+++ react-start/src/pages/layout.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
-import Header from '../components/header';
+import Header from '../components/header/header';
 
 class Layout extends React.Component {
   render() {
